requirements.txt
streamlit
groq
pymupdf
python-docx
python-dotenv